backend formatted file structure

+++ src/controllers/admin/sizeController.js
@@ -4,31 +4,35 @@ const Size = require('../../models/size');
 
 exports.createSize = async (req,res)=>{
     try{
-        if(req.body.hasOwnProperty("productId") && req.body.hasOwnProperty("name") && req.body.hasOwnProperty("status")){
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(422).json({"status": false,"message": formatError(errors.array()) });
+        }
+        else{
             let size = {
-                "name":req.body.name,
-                "status":req.body.status
+                "name":req.body.name
             }
             let newSize = new Size(size)
             await newSize.save().then((date)=>{
                 res.status(200).json({"status":true,"message":"Create Successfully"})
             })
             .catch((err)=>{
-                res.status(400).json({"status":false,"message":err})
+                res.status(400).json({"status":false,"message":err.message})
             })
-        }
-        else{
-            res.status(422).json({"status":false,"message":"Validation Error"})
         }
     }
     catch(err){
-        res.status(400).json({"status":false,"message":err})
+        res.status(400).json({"status":false,"message":err.message})
     }
 }
 
 exports.editSize = async (req,res)=>{
     try{
-        if(req.params.hasOwnProperty("id") && req.body.hasOwnProperty("name") && req.body.hasOwnProperty("status")){
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(422).json({"status": false,"message": formatError(errors.array()) });
+        }
+        else{
             let updateSize = {
                 "name": req.body.name,
                 "status": req.body.status,
@@ -36,21 +40,23 @@ exports.editSize = async (req,res)=>{
             await Size.findByIdAndUpdate(req.params.id,updateSize).then(data=>{
                 res.status(200).json({"status":true,"message":"Update Successfully"})
             })
-            .catch((error)=>{
-                res.status(400).json({"status":false,"message":error});
+            .catch((err)=>{
+                res.status(400).json({"status":false,"message":err.message});
             })
-        }
-        else{
-            res.status(422).json({"status":false,"message":"Validation Error"})
         }
     }
     catch(err){
-        res.status(400).json({"status":false,"message":err})
+        res.status(400).json({"status":false,"message":err.message})
     }
 }
 
 exports.deleteSize = async (req,res)=>{
     try{
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(422).json({"status": false,"message": formatError(errors.array()) });
+        }
+        else{
         if(req.body.hasOwnProperty("productId") && req.body.hasOwnProperty("sizeId")){
             await Category.findByIdAndDelete(req.body.sizeId).then(data=>{
                 res.status(200).json({"status":true,"message":"Deleted Successfully"})
@@ -62,7 +68,8 @@ exports.deleteSize = async (req,res)=>{
         else{
             res.status(422).json({"status":false,"message":"Validation Error"})
         }
-    }
+        }
+    }    
     catch(err){
         res.status(400).json({"status":false,"message":err})
     }
